// src/modules/payment/dto/verify-payment.dto.ts
import { ApiProperty } from "@nestjs/swagger";
import { IsNotEmpty, IsString } from "class-validator";

export class VerifyPaymentDto {
  @ApiProperty({
    description: "Razorpay Order ID returned from order creation",
    example: "order_c51d8e59e7167bc0",
  })
  @IsString()
  @IsNotEmpty()
  razorpayOrderId: string;

  @ApiProperty({
    description: "Razorpay Payment ID generated by Razorpay Checkout",
    example: "pay_xyz45678901234",
  })
  @IsString()
  @IsNotEmpty()
  razorpayPaymentId: string;

  @ApiProperty({
    description: "Cryptographic HMAC SHA-256 signature generated by Razorpay",
    example: "9ef9a616fa86c51d8e59e7167bc0...",
  })
  @IsString()
  @IsNotEmpty()
  razorpaySignature: string;
}